fix: el empaque se elige, y sirve igual para kilos y litros
Dos cosas que el cliente encontró probándolo.

La primera: el nombre del empaque era un campo de texto que venía con «Caja»
escrito. Quien no bajaba a mirarlo guardaba «Caja» sin haberlo elegido, y
después la pantalla de ingreso pedía «Cajas» de un producto que llegaba en
sacos. Ahora es una lista —caja, paquete, bolsa, saco, fardo, plancha, docena,
bidón, turril, balde, blíster— con un «Otro…» para lo que no esté, y sin valor
por defecto: hay que decirlo, y lo que se diga es lo que sale en todas las
etiquetas.

La segunda: el formulario decía que dejaras la casilla sin marcar «si el
producto se compra y se vende en la misma unidad: el arroz por kilo». Eso
mandaba justo al lado contrario a quien compra arroz en sacos de 46 kg, que es
el mismo caso que una caja de 24 gaseosas y siempre funcionó. El texto ahora
dice lo que hay: sirve igual para lo que se cuenta, lo que se pesa y lo que se
mide.

Y para que sea verdad del todo, el contenido del empaque pasa de SMALLINT a
DECIMAL(10,3). Un galón son 3.785 litros; con un entero había que redondear a
4, y ese redondeo se iba derecho al stock —10 galones cargados como 40 litros
contra los 37.85 que entraron—. El mínimo pasa de «>= 2» a «> 1», que era lo
que de verdad se quería decir y ya no deja fuera una caja de 1.5 kg.

De paso, `form/select` deja de llamar a `old(null)` cuando el select no lleva
nombre. Devolvía TODO el input anterior como array y la comparación reventaba
con «Array to string conversion», tumbando la pantalla entera con un 500. El
`form/input` ya se guardaba de esto; el select no.

+++ sistema-ventas/resources/views/inventario/index.blade.php
@@ -142,7 +142,7 @@
                                     'paso' => $unidad?->permite_decimal ? 0.001 : 1,
                                     'proveedor' => $producto->proveedor_id,
                                     'compra' => (float) $producto->precio_compra,
-                                    'contenido' => (int) $producto->contenido_empaque,
+                                    'contenido' => (float) $producto->contenido_empaque,
                                     'empaque' => mb_strtolower($producto->nombre_empaque ?? ''),
                                     'desglose' => $producto->stock_desglosado,
                                 ];
